+++ includes/admin/class-wp-statistics-admin-assets.php
@@ -195,7 +195,7 @@ class Admin_Assets
         }
 
         //Load Jquery VMap Css
-        if (!Option::get('disable_map') and (Menus::in_page('overview') || Menus::in_page('pages') || (in_array($screen_id, array('dashboard')) and !Option::get('disable_dashboard')))) {
+        if (Menus::in_page('overview') || Menus::in_page('pages') || (in_array($screen_id, array('dashboard')) and !Option::get('disable_dashboard'))) {
             wp_enqueue_style(self::$prefix . '-jqvmap', self::url('jqvmap/jqvmap.min.css'), array(), '1.5.1');
         }
 
@@ -245,7 +245,7 @@ class Admin_Assets
         }
 
         // Load Jquery VMap Js Library
-        if (!Option::get('disable_map') and (Menus::in_page('overview') || Menus::in_page('pages') || (in_array($screen_id, array('dashboard')) and !Option::get('disable_dashboard')))) {
+        if (Menus::in_page('overview') || Menus::in_page('pages') || (in_array($screen_id, array('dashboard')) and !Option::get('disable_dashboard'))) {
             wp_enqueue_script(self::$prefix . '-jqvmap', self::url('jqvmap/jquery.vmap.min.js'), array('jquery'), "1.5.1", ['in_footer' => true]);
             wp_enqueue_script(self::$prefix . '-jqvmap-world', self::url('jqvmap/jquery.vmap.world.min.js'), array('jquery'), "1.5.1", ['in_footer' => true]);
         }
@@ -411,7 +411,7 @@ class Admin_Assets
             'er_valid_ip'                  => __('Enter a Valid IP Address', 'wp-statistics'),
             'please_wait'                  => __('Loading, Please Wait...', 'wp-statistics'),
             'user'                         => __('User Information', 'wp-statistics'),
-            'rest_connect'                 => __('Error connecting to WordPress REST API. Disable ad-blocker for this page or unblock /wp-json/wp-statistics/v2/metabox in the ad-blocker configuration.', 'wp-statistics'),
+            'rest_connect'                 => __('Failed to retrieve data. Please check the browser console and the XHR request under Network → XHR for details.', 'wp-statistics'),
             'privacy_compliant'            => __('Your WP Statistics settings are privacy-compliant.', 'wp-statistics'),
             'non_privacy_compliant'        => __('Your WP Statistics settings are not privacy-compliant. Please update your settings.', 'wp-statistics'),
             'privacy_resolve_alert'        => __('By manually resolving this item, please ensure your website’s privacy policy is updated to accurately reflect this setting. This is essential for maintaining compliance and transparency with your users.', 'wp-statistics'),
@@ -427,7 +427,7 @@ class Admin_Assets
         $list['active_post_type'] = Helper::getPostTypeName(Request::get('pt', 'post'));
 
         // Rest-API Meta Box Url
-        $list['stats_report_option']    = Option::get('stats_report') ? true : false;
+        $list['stats_report_option']    = Option::get('time_report') == '0' ? false : true;
         $list['setting_url']            = Menus::admin_url('settings');
         $list['admin_url']              = admin_url();
         $list['ajax_url']               = admin_url('admin-ajax.php');