+++ includes/admin/templates/pages/posts/posts-report-locked.php
@@ -1,8 +1,0 @@
-<?php
-
-use WP_STATISTICS\Admin_Template;
-
-Admin_Template::get_template(['layout/partials/data-plus-locked-page'], [
-    'campaign'  => 'content-analytics',
-    'src'       => 'assets/images/locked/posts-report.jpg',
-]);