+++ includes/api/v2/class-wp-statistics-api-hit.php
@@ -2,6 +2,7 @@
 
 namespace WP_STATISTICS\Api\v2;
 
+use Exception;
 use WP_STATISTICS\Hits;
 
 class Hit extends \WP_STATISTICS\RestAPI
@@ -65,21 +66,30 @@ class Hit extends \WP_STATISTICS\RestAPI
      * Record WP Statistics when Cache is enable
      *
      * @return \WP_REST_Response
-     * @throws \Exception
+     * @throws Exception
      */
     public function hit_callback()
     {
-        // Start Record
-        $exclusion    = Hits::record();
-        $responseData = [
-            'status' => $exclusion['exclusion_match'] == false,
-        ];
+        $statusCode = false;
 
-        if ($exclusion['exclusion_match']) {
-            $responseData['data'] = $exclusion;
+        try {
+            Hits::record();
+            $responseData['status'] = true;
+
+        } catch (Exception $e) {
+            $responseData['status'] = false;
+            $responseData['data']   = $e->getMessage();
+            $statusCode             = $e->getCode();
         }
 
         $response = rest_ensure_response($responseData);
+
+        /**
+         * Set the status code
+         */
+        if ($statusCode) {
+            $response->set_status($statusCode);
+        }
 
         /**
          * Set headers for the response