+++ includes/class-wp-statistics-exclusion.php
@@ -414,16 +414,18 @@ class Exclusion
         }
 
         // Check user ip is empty or not user agent
-        if (Option::get('corrupt_browser_info')) {
-            if ($visitorProfile->getHttpUserAgent() == '' || $visitorProfile->getIp() == '') {
-                return true;
-            }
+        if ($visitorProfile->getHttpUserAgent() == '' || $visitorProfile->getIp() == '') {
+            return true;
+        }
 
-            $userAgent = $visitorProfile->getUserAgent();
+        $userAgent = $visitorProfile->getUserAgent();
 
-            if (!$userAgent['isBrowserDetected'] && !$userAgent['isPlatformDetected']) {
-                return true;
-            }
+        if (isset($userAgent['isBot']) && $userAgent['isBot'] === true) {
+            return true;
+        }
+
+        if (!$userAgent['isBrowserDetected'] && !$userAgent['isPlatformDetected']) {
+            return true;
         }
 
         return false;