+++ includes/class-wp-statistics-frontend.php
@@ -67,10 +67,11 @@ class Frontend
             'option'               => [
                 'consentLevel'         => Option::get('consent_level_integration', 'disabled'),
                 'dntEnabled'           => Option::get('do_not_track'),
-                'cacheCompatibility'   => Option::get('use_cache_plugin'),
+                'isClientSideTracking' => Option::get('use_cache_plugin'),
                 'isWpConsentApiActive' => WpConsentApi::isWpConsentApiActive(),
                 'trackAnonymously'     => Helper::shouldTrackAnonymously()
             ],
+            'jsCheckTime'          => apply_filters('wp_statistics_js_check_time_interval', 60000)
         );
 
         Assets::script('tracker', 'js/tracker.js', [], $jsArgs, true, Option::get('bypass_ad_blockers', false));