+++ includes/class-wp-statistics-hits.php
@@ -2,6 +2,7 @@
 
 namespace WP_STATISTICS;
 
+use Exception;
 use WP_Statistics\Components\Singleton;
 use WP_Statistics\Service\Analytics\VisitorProfile;
 use WP_Statistics\Service\Integrations\WpConsentApi;
@@ -25,30 +26,30 @@ class Hits extends Singleton
     /**
      * WP_Statistics Hits Class.
      *
-     * @throws \Exception
+     * @throws Exception
      */
     public function __construct()
     {
 
-        # Sanitize Hit Data if Has Rest-Api Process
+        // Sanitize Hit Data if Has Rest-Api Process
         if (self::is_rest_hit()) {
 
-            # Get Hit Data
+            // Get Hit Data
             $this->rest_hits = (object)self::rest_params();
 
-            # Filter Data
+            // Filter Data
             add_filter('wp_statistics_current_page', array($this, 'set_current_page'));
             add_filter('wp_statistics_page_uri', array($this, 'set_page_uri'));
             add_filter('wp_statistics_user_id', array($this, 'set_current_user'));
         }
 
-        # Record Login Page Hits
+        // Record Login Page Hits
         if (!Option::get('exclude_loginpage')) {
-            add_action('init', array($this, 'record_login_page_hits'));
+            add_action('init', array($this, 'trackLoginPageCallback'));
         }
 
-        # Record WordPress Front Page Hits
-        add_action('wp', array($this, 'record_wp_hits'));
+        // Server Side Tracking
+        add_action('wp', array($this, 'trackServerSideCallback'));
     }
 
     /**
@@ -130,9 +131,9 @@ class Hits extends Singleton
     }
 
     /**
-     * Get Visitor information and Record To DB
+     * Record the visitor
      *
-     * @throws \Exception
+     * @throws Exception
      */
     public static function record($visitorProfile = null)
     {
@@ -140,46 +141,102 @@ class Hits extends Singleton
             $visitorProfile = new VisitorProfile();
         }
 
-        # Check Exclusion This Hits
+        /**
+         * Check the exclusion
+         */
         $exclusion = Exclusion::check($visitorProfile);
 
-        # Record Hits Exclusion
+        /**
+         * Record exclusion if needed & then skip the tracking
+         */
         if ($exclusion['exclusion_match'] === true) {
             Exclusion::record($exclusion);
+
+            throw new Exception($exclusion['exclusion_reason'], 403);
         }
 
-        # Record User Views
-        if (Visit::active() and $exclusion['exclusion_match'] === false) {
+        /**
+         * Record User Views
+         */
+        if (Visit::active()) {
             Visit::record();
         }
 
-        # Record Visitor Detail
+        /**
+         * Record Visitor Detail
+         */
+        $visitorId = false;
         if (Visitor::active()) {
-            $visitor_id = Visitor::record($visitorProfile, $exclusion);
+            $visitorId = Visitor::record($visitorProfile);
         }
 
-        # Record Search Engine
-        if (isset($visitor_id) and $visitor_id > 0 and $exclusion['exclusion_match'] === false) {
-            SearchEngine::record(array('visitor_id' => $visitor_id));
+        /**
+         * Record Search Engine
+         */
+        if ($visitorId) {
+            SearchEngine::record(array('visitor_id' => $visitorId));
         }
 
-        # Record Pages
-        if (Pages::active() and $exclusion['exclusion_match'] === false) {
-            $page_id = Pages::record($visitorProfile);
+        /**
+         * Record Pages
+         */
+        $pageId = false;
+        if (Pages::active()) {
+            $pageId = Pages::record($visitorProfile);
         }
 
-        # Record Visitor Relationship
-        if (isset($visitor_id) and $visitor_id > 0 and isset($page_id) and $page_id > 0) {
-            Visitor::save_visitors_relationships($page_id, $visitor_id);
+        /**
+         * Record Visitor Relationship
+         */
+        if ($visitorId && $pageId) {
+            Visitor::save_visitors_relationships($pageId, $visitorId);
         }
 
-        # Record User Online
-        if (UserOnline::active() and ($exclusion['exclusion_match'] === false)) {
-            $pageID = isset($page_id) && $page_id > 0 ? $page_id : 0;
-            UserOnline::record($visitorProfile, [
-                'page_id' => $pageID,
-            ]);
+        /**
+         * Record User Online with the visitor request in the same time.
+         */
+        self::recordOnline($visitorProfile, $exclusion, $pageId);
+
+        return $exclusion;
+    }
+
+    /**
+     * Record the user online standalone
+     *
+     * @throws Exception
+     */
+    public static function recordOnline($visitorProfile = null, $exclusion = null, $pageId = null)
+    {
+        if (!UserOnline::active()) {
+            return;
         }
+
+        if (!$visitorProfile) {
+            $visitorProfile = new VisitorProfile();
+        }
+
+        /**
+         * Check the exclusion
+         */
+        if (!$exclusion) {
+            $exclusion = Exclusion::check($visitorProfile);
+        }
+
+        /**
+         * Record exclusion if needed & then skip the tracking
+         */
+        if ($exclusion['exclusion_match'] === true) {
+            Exclusion::record($exclusion);
+
+            throw new Exception($exclusion['exclusion_reason'], 403);
+        }
+
+        $args = null;
+        if ($pageId) {
+            $args['page_id'] = $pageId;
+        }
+
+        UserOnline::record($visitorProfile, $args);
 
         return $exclusion;
     }
@@ -187,30 +244,39 @@ class Hits extends Singleton
     /**
      * Record Hits in Login Page
      *
-     * @throws \Exception
+     * @throws Exception
      */
-    public static function record_login_page_hits()
+    public static function trackLoginPageCallback()
     {
         if (Helper::is_login_page()) {
-            self::record();
+            try {
+                self::record();
+            } catch (Exception $e) {
+
+            }
         }
     }
 
     /**
-     * Record WordPress Frontend Hits
+     * Server-Side Tracking Callback
      *
-     * @throws \Exception
+     * @throws Exception
      */
-    public static function record_wp_hits()
+    public static function trackServerSideCallback()
     {
-        if (is_admin() or is_preview() or Option::get('use_cache_plugin') or Helper::dntEnabled()) {
-            return;
-        }
+        try {
+            if (is_admin() or is_preview() or Option::get('use_cache_plugin') or Helper::dntEnabled()) {
+                return;
+            }
 
-        $consentLevel = Option::get('consent_level_integration', 'disabled');
+            $consentLevel = Option::get('consent_level_integration', 'disabled');
 
-        if ($consentLevel == 'disabled' || Helper::shouldTrackAnonymously() || !WpConsentApi::isWpConsentApiActive() || !function_exists('wp_has_consent') || wp_has_consent($consentLevel)) {
-            self::record();
+            if ($consentLevel == 'disabled' || Helper::shouldTrackAnonymously() || !WpConsentApi::isWpConsentApiActive() || !function_exists('wp_has_consent') || wp_has_consent($consentLevel)) {
+                self::record();
+            }
+
+        } catch (Exception $e) {
+
         }
     }
 }